requirements.txt
beautifulsoup4
Django
django-choices~=1.7.1
djangorestframework
enlighten
gunicorn==20.1.0
httpx
python-dotenv
requests
django-filter
drf-yasg
elasticsearch==7.13.1
gspread
tqdm
bs4
django-import-export
psycopg2~=2.9.1

aiohttp
asyncpg
pandas
APScheduler
django-apscheduler
tortoise-orm

Pillow
phonenumbers
beautifulsoup4
python-dateutil
django-extensions

djantimat
XlsxWriter
emoji
kafka-python
celery